dataset: Add discriminator dataset
The way this one works is that we sample two goals,
then two images for the first goal (anchor and true)
and one image from the second goal (false). Then we can use
this in a kind of contrastive-learning process, where we try
to maximize agreement between anchor and true and try to maximize
disagreement between anchor and false.