models: Add "IndependentAttention" model
In this model we treat the components separately from each other
and then use log-sum-exp in order to "multiply" the results together
(eg, both attributes must be matched indepdendently).

In combination with L1 regularization on the attention maps, we
can ensure that we explain the problem sparsely and get a more generalizable
representation with few samples.