scripts: Create the parallel_env before loading the dataset
The dataset is 4GB. If we load it before creating the parallel
env, we have to account for all that memory when fork()ing
off the child processes. Even though linux is copy-on-write,
its a huge liability - if we manage to even touch a reference
on that memory, entire pages get copied into the child process
which balloons out memory usage. To quarantine the children from
this, we create them before loading all that data into memory.