models: add VIN based models
The VIN model is based on MVProp, which is an explicit planning module,
treating the image as a graph of states where value can propagate
from one state to another. The task is to learn the propagation
function. In our formulation, we don't know where the agent is, so
we combine the thing with one big convnet in order to determine what
the true values are.