discriminator: Sample a null-goal state with 1/(n_goals - 1 + 1) probability
The "destination" might sometimes be a null goal, meaning
that instead of comparing to a different goal, we compare to a non-goal
state.

We have probability 1/(n_goals - 1 + 1) so that there is a an equal
probability of comparing to any other goal, or a null goal for the
second goal (since we do not sample the first sampled goal
the first time around).